genbank API interface to run
